Refactor to desired structure (#93)
* Refactor to desired structure

* fix metric test

* Update branch.yaml

* cargo fmt

* fix feature flags

* simplify feature flags for now

* add env_dependency
