+++ dsh_sdk/Cargo.toml
@@ -17,13 +17,12 @@ all-features = true
 [dependencies]
 base64 = {version = "0.22", optional = true }
 bytes = { version = "1.6", optional = true }
-dashmap = {version = "6.0", optional = true}
 http-body-util = { version = "0.1", optional = true }
 hyper = { version = "1.3", features = ["server", "http1"], optional = true }
 hyper-util = { version = "0.1", features = ["tokio"], optional = true }
 lazy_static = { version = "1.5", optional = true }
 log = "0.4"
-pem = "3"
+pem = {version = "3", optional = true }
 prometheus = { version = "0.13", features = ["process"], optional = true }
 rcgen = { version = "0.13", optional = true }
 reqwest = { version = "0.12", default-features = false, features = ["rustls-tls", "json", "blocking"], optional = true }
@@ -36,22 +35,29 @@ tokio = { version = "^1.35", features = ["signal", "sync", "time", "macros"], op
 tokio-util = { version = "0.7", default-features = false, optional = true }
 
 [features]
-default = ["bootstrap", "graceful_shutdown", "metrics", "rdkafka-ssl"]
+default = ["bootstrap", "graceful-shutdown", "metrics", "rdkafka-ssl"]
 
-bootstrap = ["rcgen", "serde_json", "reqwest", "tokio/rt-multi-thread"]
+bootstrap = ["certificate", "serde_json", "tokio/rt-multi-thread"] 
+certificate = ["rcgen", "reqwest", "pem"]
 metrics = ["prometheus", "hyper", "hyper-util", "http-body-util", "lazy_static", "tokio", "bytes"]
-dlq = ["tokio", "bootstrap", "rdkafka-ssl", "graceful_shutdown"]
-graceful_shutdown = ["tokio", "tokio-util"]
-rdkafka-ssl-vendored = ["rdkafka", "rdkafka/ssl-vendored"]
-rdkafka-ssl = ["rdkafka", "rdkafka/ssl"]
-rest-token-fetcher = ["reqwest"] 
-mqtt-token-fetcher = ["base64","dashmap","reqwest","serde_json","sha2","tokio/sync"]
+dlq = ["tokio", "bootstrap", "rdkafka-ssl", "graceful-shutdown"]
+graceful-shutdown = ["tokio", "tokio-util"]
+management-api = ["reqwest"]
+protocol-token-fetcher = ["base64", "reqwest", "serde_json", "sha2", "tokio/sync"]
+# http-protocol-adapter = ["protocol-token-fetcher"]
+# mqtt-protocol-adapter = ["protocol-token-fetcher"]
+
+# TODO: Remove the following features at v0.6.0
+rdkafka-ssl-vendored = ["rdkafka", "rdkafka/ssl-vendored", "rdkafka/cmake-build"]
+rdkafka-ssl = ["rdkafka", "rdkafka/ssl", "rdkafka/cmake-build"]
+#mqtt-token-fetcher = ["protocol-token-fetcher"]
+#rest-token-fetcher = ["management-api"] 
 
 [dev-dependencies]
 mockito = "1.1.1"
 openssl = "0.10"
 tokio = { version = "^1.35", features = ["full"] }
-hyper = { version = "1.2.0", features = ["full"]}
+hyper = { version = "1.3", features = ["full"]}
 serial_test = "3.1.0"
 dsh_rest_api_client = { path = "../dsh_rest_api_client", version = "0.2.0"}
 dsh_sdk = { features = ["dlq"], path = "." }
