+++ dsh_sdk/examples/mqtt_token_fetcher.rs
@@ -6,8 +6,7 @@ use dsh_sdk::mqtt_token_fetcher::{MqttToken, MqttTokenFetcher};
 async fn main() {
     let tenant_name = env::var("TENANT").unwrap().to_string();
     let api_key = env::var("API_KEY").unwrap().to_string();
-    let mqtt_token_fetcher: MqttTokenFetcher =
-        MqttTokenFetcher::new(tenant_name, api_key, dsh_sdk::Platform::NpLz);
+    let mqtt_token_fetcher = MqttTokenFetcher::new(tenant_name, api_key, dsh_sdk::Platform::NpLz);
     let token: MqttToken = mqtt_token_fetcher
         .get_token("Client-id", None) //Claims = None fetches all possible claims
         .await
