+++ dsh_sdk/examples/mqtt_token_fetcher_specific_claims.rs
@@ -16,8 +16,7 @@ async fn main() {
 
     let claims_vector = vec![claims_sub, claims_pub];
 
-    let mqtt_token_fetcher: MqttTokenFetcher =
-        MqttTokenFetcher::new(tenant_name, api_key, dsh_sdk::Platform::NpLz);
+    let mqtt_token_fetcher = MqttTokenFetcher::new(tenant_name, api_key, dsh_sdk::Platform::NpLz);
 
     let token: MqttToken = mqtt_token_fetcher
         .get_token("Client-id", Some(claims_vector))
