+++ dsh_sdk/src/certificates/pki_config_dir.rs
@@ -5,7 +5,7 @@
 //!
 //! This also makes it possible to use the DSH SDK with Kafka Proxy
 //! or VPN outside of the DSH environment.
-use super::certificates::Cert;
+use super::Cert;
 use crate::error::DshError;
 use crate::{utils, VAR_PKI_CONFIG_DIR};
 
@@ -14,8 +14,13 @@ use pem::{self, Pem};
 use rcgen::KeyPair;
 use std::path::{Path, PathBuf};
 
-pub(crate) fn get_pki_cert() -> Result<Cert, DshError> {
-    let config_dir = PathBuf::from(utils::get_env_var(VAR_PKI_CONFIG_DIR)?);
+pub(crate) fn get_pki_certificates<P>(pki_config_dir: Option<P>) -> Result<Cert, DshError>
+where
+    P: AsRef<Path>,
+{
+    let config_dir = pki_config_dir
+        .map(|dir| dir.as_ref().to_path_buf())
+        .unwrap_or(PathBuf::from(utils::get_env_var(VAR_PKI_CONFIG_DIR)?));
     let ca_cert_paths = get_file_path_bufs("ca", PkiFileType::Cert, &config_dir)?;
     let dsh_ca_certificate_pem = get_certificate(ca_cert_paths)?;
     let client_cert_paths = get_file_path_bufs("client", PkiFileType::Cert, &config_dir)?;
@@ -33,8 +38,8 @@ pub(crate) fn get_pki_cert() -> Result<Cert, DshError> {
 /// Get certificate from the PKI config directory
 ///
 /// Looks for all files containing client*.pem and client.crt in the PKI config directory.
-fn get_certificate(mut cert_paths: Vec<PathBuf>) -> Result<Vec<Pem>, DshError> {
-    'file: while let Some(file) = cert_paths.pop() {
+fn get_certificate(cert_paths: Vec<PathBuf>) -> Result<Vec<Pem>, DshError> {
+    'file: for file in cert_paths {
         info!("{} - Reading certificate file", file.display());
         if let Ok(ca_cert) = std::fs::read(&file) {
             let pem_result = pem::parse_many(&ca_cert);
@@ -61,29 +66,33 @@ fn get_certificate(mut cert_paths: Vec<PathBuf>) -> Result<Vec<Pem>, DshError> {
     Err(DshError::NoCertificates)
 }
 
-/// Get certificate from the PKI config directory
+/// Get key pair from a file in the PKI config directory
 ///
-/// Looks for all files containing client*.pem and client.crt in the PKI config directory.
-fn get_key_pair(mut key_paths: Vec<PathBuf>) -> Result<KeyPair, DshError> {
-    while let Some(file) = key_paths.pop() {
+/// Returns first succesfull converted key pair found in the given list of paths
+fn get_key_pair(key_paths: Vec<PathBuf>) -> Result<KeyPair, DshError> {
+    for file in key_paths {
         info!("{} - Reading key file", file.display());
         if let Ok(bytes) = std::fs::read(&file) {
-            if let Ok(string) = std::string::String::from_utf8(bytes) {
+            if let Ok(string) = std::string::String::from_utf8(bytes.clone()) {
                 debug!("{} - Key parsed as string", file.display());
-                match rcgen::KeyPair::from_pem(&string) {
-                    Ok(key_pair) => {
-                        debug!("{} - Key parsed as KeyPair from string", file.display());
-                        return Ok(key_pair);
-                    }
-                    Err(e) => warn!("{} - Error parsing key: {:?}", file.display(), e),
+                if let Ok(key_pair) = rcgen::KeyPair::from_pem(&string) {
+                    debug!("{} - Key parsed as KeyPair from string", file.display());
+                    return Ok(key_pair);
+                } else {
+                    warn!("{} - Error parsing key from string", file.display());
                 }
+            }
+            if let Ok(key_pair) = rcgen::KeyPair::try_from(bytes) {
+                debug!("{} - Key parsed as KeyPair from bytes", file.display());
+                return Ok(key_pair);
+            } else {
+                warn!("{} - Error parsing key from bytes", file.display());
             }
         }
     }
     info!("No (valid) key found in the PKI config directory");
     Err(DshError::NoCertificates)
 }
-
 /// Get the path to the PKI config direc
 fn get_file_path_bufs<P>(
     prefix: &str,
@@ -136,22 +145,26 @@ mod tests {
     use serial_test::serial;
 
     const PKI_CONFIG_DIR: &str = "test_files/pki_config_dir";
-    const PKI_KEY_FILE_NAME: &str = "client.key";
+    const PKI_KEY_FILE_PEM_NAME: &str = "client.key";
+    const PKI_KEY_FILE_DER_NAME: &str = "client-der.key";
     const PKI_CERT_FILE_NAME: &str = "client.pem";
     const PKI_CA_FILE_NAME: &str = "ca.crt";
 
     fn create_test_pki_config_dir() {
         let path = PathBuf::from(PKI_CONFIG_DIR);
-        let path_key = PathBuf::from(PKI_CONFIG_DIR).join(PKI_KEY_FILE_NAME);
+        let path_key_pem = PathBuf::from(PKI_CONFIG_DIR).join(PKI_KEY_FILE_PEM_NAME);
+        let path_key_der = PathBuf::from(PKI_CONFIG_DIR).join(PKI_KEY_FILE_DER_NAME);
         let path_cert = PathBuf::from(PKI_CONFIG_DIR).join(PKI_CERT_FILE_NAME);
         let path_ca = PathBuf::from(PKI_CONFIG_DIR).join(PKI_CA_FILE_NAME);
-        if path_key.exists() && path_cert.exists() && path_ca.exists() {
+        if path_key_pem.exists() && path_cert.exists() && path_ca.exists() && path_key_der.exists()
+        {
             return;
         }
         let _ = std::fs::create_dir(path);
         let priv_key = openssl::rsa::Rsa::generate(2048).unwrap();
         let pkey = PKey::from_rsa(priv_key).unwrap();
-        let key = pkey.private_key_to_pem_pkcs8().unwrap();
+        let key_pem = pkey.private_key_to_pem_pkcs8().unwrap();
+        let key_der = pkey.private_key_to_pkcs8().unwrap();
         let mut x509_name = openssl::x509::X509NameBuilder::new().unwrap();
         x509_name.append_entry_by_text("CN", "test_ca").unwrap();
         let x509_name = x509_name.build();
@@ -168,7 +181,8 @@ mod tests {
         let x509 = x509.build();
         let ca_cert = x509.to_pem().unwrap();
         let cert = x509.to_pem().unwrap();
-        std::fs::write(path_key, key).unwrap();
+        std::fs::write(path_key_pem, key_pem).unwrap();
+        std::fs::write(path_key_der, key_der).unwrap();
         std::fs::write(path_ca, ca_cert).unwrap();
         std::fs::write(path_cert, cert).unwrap();
     }
@@ -181,7 +195,7 @@ mod tests {
         let result_cert = get_file_path_bufs("client", PkiFileType::Cert, &path).unwrap();
         assert_eq!(result_cert.len(), 1);
         let result_key = get_file_path_bufs("client", PkiFileType::Key, &path).unwrap();
-        assert_eq!(result_key.len(), 1);
+        assert_eq!(result_key.len(), 2);
         assert_ne!(result_cert, result_key);
         let result_ca = get_file_path_bufs("ca", PkiFileType::Cert, &path).unwrap();
         assert_eq!(result_ca.len(), 1);
@@ -194,7 +208,7 @@ mod tests {
     #[serial(pki)]
     fn test_get_certificate() {
         create_test_pki_config_dir();
-        let path_key = PathBuf::from(PKI_CONFIG_DIR).join(PKI_KEY_FILE_NAME);
+        let path_key = PathBuf::from(PKI_CONFIG_DIR).join(PKI_KEY_FILE_PEM_NAME);
         let path_cert = PathBuf::from(PKI_CONFIG_DIR).join(PKI_CERT_FILE_NAME);
         let path_ca = PathBuf::from(PKI_CONFIG_DIR).join(PKI_CA_FILE_NAME);
         let path_ne = PathBuf::from(PKI_CONFIG_DIR).join("not_existing.crt");
@@ -223,9 +237,30 @@ mod tests {
 
     #[test]
     #[serial(pki)]
-    fn test_get_key_pair() {
+    fn test_get_key_pair_pem() {
         create_test_pki_config_dir();
-        let path_key = PathBuf::from(PKI_CONFIG_DIR).join(PKI_KEY_FILE_NAME);
+        let path_key = PathBuf::from(PKI_CONFIG_DIR).join(PKI_KEY_FILE_PEM_NAME);
+        let path_cert = PathBuf::from(PKI_CONFIG_DIR).join(PKI_CERT_FILE_NAME);
+        let path_ca = PathBuf::from(PKI_CONFIG_DIR).join(PKI_CA_FILE_NAME);
+        let path_ne = PathBuf::from(PKI_CONFIG_DIR).join("not_existing.key");
+        let result = get_key_pair(vec![path_key.clone()]);
+        assert!(result.is_ok());
+        let result = get_key_pair(vec![path_ne.clone(), path_key.clone()]);
+        assert!(result.is_ok());
+        let result =
+            get_key_pair(vec![path_ne.clone(), path_cert.clone(), path_ca.clone()]).unwrap_err();
+        assert!(matches!(result, DshError::NoCertificates));
+        let result = get_key_pair(vec![]).unwrap_err();
+        assert!(matches!(result, DshError::NoCertificates));
+        let result = get_key_pair(vec![path_ne]).unwrap_err();
+        assert!(matches!(result, DshError::NoCertificates));
+    }
+
+    #[test]
+    #[serial(pki)]
+    fn test_get_key_pair_der() {
+        create_test_pki_config_dir();
+        let path_key = PathBuf::from(PKI_CONFIG_DIR).join(PKI_KEY_FILE_DER_NAME);
         let path_cert = PathBuf::from(PKI_CONFIG_DIR).join(PKI_CERT_FILE_NAME);
         let path_ca = PathBuf::from(PKI_CONFIG_DIR).join(PKI_CA_FILE_NAME);
         let path_ne = PathBuf::from(PKI_CONFIG_DIR).join("not_existing.key");
@@ -246,10 +281,10 @@ mod tests {
     #[serial(pki, env_dependency)]
     fn test_get_pki_cert() {
         create_test_pki_config_dir();
-        let result = get_pki_cert().unwrap_err();
-        assert!(matches!(result, DshError::EnvVarError(_)));
+        let result = get_pki_certificates::<PathBuf>(None).unwrap_err();
+        assert!(matches!(result, DshError::EnvVarError(_, _)));
         std::env::set_var(VAR_PKI_CONFIG_DIR, PKI_CONFIG_DIR);
-        let result = get_pki_cert();
+        let result = get_pki_certificates::<PathBuf>(None);
         assert!(result.is_ok());
         std::env::remove_var(VAR_PKI_CONFIG_DIR);
     }
