+++ dsh_sdk/src/dsh_old/mod.rs
@@ -21,13 +21,36 @@
 //! # Ok(())
 //! # }
 //! ```
-mod bootstrap;
+
+#[deprecated(
+    since = "0.5.0",
+    note = "`dsh_sdk::dsh::certificates` is moved to `dsh_sdk::certificates`"
+)]
 pub mod certificates;
-mod config;
+#[deprecated(
+    since = "0.5.0",
+    note = "`dsh_sdk::dsh::datastream` is moved to `dsh_sdk::datastream`"
+)]
 pub mod datastream;
-mod pki_config_dir;
+#[deprecated(
+    since = "0.5.0",
+    note = "`dsh_sdk::dsh::properties` is moved to `dsh_sdk::dsh`"
+)]
 pub mod properties;
 
 // Re-export the properties struct to avoid braking changes
-pub use super::utils::get_configured_topics;
+
+#[deprecated(
+    since = "0.5.0",
+    note = "get_configured_topics is moved to `dsh_sdk::utils::get_configured_topics`"
+)]
+pub fn get_configured_topics() -> Result<Vec<String>, crate::error::DshError> {
+    let kafka_topic_string = crate::utils::get_env_var("TOPICS")?;
+    Ok(kafka_topic_string
+        .split(',')
+        .map(str::trim)
+        .map(String::from)
+        .collect())
+}
+
 pub use properties::Properties;
