+++ dsh_sdk/src/error.rs
@@ -7,11 +7,11 @@ use thiserror::Error;
 pub enum DshError {
     #[error("IO Error: {0}")]
     IoError(#[from] std::io::Error),
-    #[error("Env var error: {0}")]
-    EnvVarError(#[from] std::env::VarError),
+    #[error("Env variable {0} error: {1}")]
+    EnvVarError(String, std::env::VarError),
     #[error("Convert bytes to utf8 error: {0}")]
     Utf8(#[from] std::string::FromUtf8Error),
-    #[cfg(any(feature = "bootstrap", feature = "mqtt-token-fetcher"))]
+    #[cfg(any(feature = "bootstrap", feature = "protocol-token-fetcher"))]
     #[error("Error calling: {url}, status code: {status_code}, error body: {error_body}")]
     DshCallError {
         url: String,
@@ -21,24 +21,24 @@ pub enum DshError {
     #[cfg(feature = "bootstrap")]
     #[error("Certificates are not set")]
     NoCertificates,
-    #[cfg(feature = "bootstrap")]
+    #[cfg(any(feature = "bootstrap", feature = "pki-config-dir"))]
     #[error("Invalid PEM certificate: {0}")]
     PemError(#[from] pem::PemError),
-    #[cfg(any(feature = "bootstrap", feature = "mqtt-token-fetcher"))]
+    #[cfg(any(feature = "certificate", feature = "protocol-token-fetcher"))]
     #[error("Reqwest: {0}")]
     ReqwestError(#[from] reqwest::Error),
-    #[cfg(any(feature = "bootstrap", feature = "mqtt-token-fetcher"))]
+    #[cfg(any(feature = "bootstrap", feature = "protocol-token-fetcher"))]
     #[error("Serde_json error: {0}")]
     JsonError(#[from] serde_json::Error),
     #[cfg(feature = "bootstrap")]
     #[error("Rcgen error: {0}")]
     PrivateKeyError(#[from] rcgen::Error),
-    #[cfg(any(feature = "bootstrap", feature = "mqtt-token-fetcher"))]
+    #[cfg(any(feature = "bootstrap", feature = "protocol-token-fetcher"))]
     #[error("Error parsing: {0}")]
     ParseDnError(String),
     #[cfg(feature = "bootstrap")]
     #[error("Error getting group id, index out of bounds for {0}")]
-    IndexGroupIdError(crate::dsh::datastream::GroupType),
+    IndexGroupIdError(crate::datastream::GroupType),
     #[error("No tenant name found")]
     NoTenantName,
     #[cfg(feature = "bootstrap")]
@@ -46,7 +46,7 @@ pub enum DshError {
     NotFoundTopicError(String),
     #[cfg(feature = "bootstrap")]
     #[error("Error in topic permissions: {0} does not have {1:?} permissions.")]
-    TopicPermissionsError(String, crate::dsh::datastream::ReadWriteAccess),
+    TopicPermissionsError(String, crate::datastream::ReadWriteAccess),
     #[cfg(feature = "metrics")]
     #[error("Prometheus error: {0}")]
     Prometheus(#[from] prometheus::Error),
@@ -55,7 +55,7 @@ pub enum DshError {
     HyperError(#[from] hyper::http::Error),
 }
 
-#[cfg(feature = "rest-token-fetcher")]
+#[cfg(feature = "management-api")]
 #[derive(Error, Debug)]
 #[non_exhaustive]
 pub enum DshRestTokenError {
