+++ dsh_sdk/src/protocol_adapters/http_protocol/mod.rs
@@ -1,0 +1,3 @@
+//! DSH Http Protocol client
+//!
+//! To be created
