+++ dsh_sdk/src/protocol_adapters/mod.rs
@@ -1,0 +1,11 @@
+#[cfg(feature = "http-protocol-adapter")]
+pub mod http_protocol;
+pub mod kafka_protocol;
+#[cfg(feature = "mqtt-protocol-adapter")]
+pub mod mqtt_protocol;
+#[cfg(feature = "protocol-token-fetcher")]
+pub mod token_fetcher;
+
+#[cfg(feature = "protocol-token-fetcher")]
+#[doc(inline)]
+pub use token_fetcher::ProtocolTokenFetcher;
