+++ dsh_sdk/src/protocol_adapters/mqtt_protocol/mod.rs
@@ -1,0 +1,3 @@
+//! DSH Mqtt Protocol client
+//!
+//! To be created
